packages/fantasion-backend/pyproject.toml
[tool.poetry]
name = "fantasion-backend"
version = "0.1.0"
description = "Backend for Fantasion website"
authors = ["Pavel Žák <pavel@zak.global>"]
packages = [
  { include = "fantasion" },
  { include = "fantasion_api" },
  { include = "fantasion_banking" },
  { include = "fantasion_content" },
  { include = "fantasion_domain" },
  { include = "fantasion_eshop" },
  { include = "fantasion_expeditions" },
  { include = "fantasion_generics" },
  { include = "fantasion_locations" },
  { include = "fantasion_people" },
  { include = "fantasion_signups" },
]
include = [
  "manage.py",
  "**/*.mo",
]

[tool.poetry.dependencies]
python = "^3.9"
Django = "^3.2.12"
django-extensions = "^3.1.5"
django-modeltranslation = "^0.17.3"
django-nested-admin = "^3.3.3"
czech-sort = "^1.0.1"
Pillow = "^9.0.1"
ffmpeg = "^1.4"
djangorestframework = "^3.13.1"
Markdown = "^3.3.6"
django-filter = "^21.1"
gunicorn = "^20.1.0"
django-cors-headers = "^3.10.1"
psycopg2-binary = "^2.9.3"
django-storages = {extras = ["google"], version = "^1.12.3"}
djangorestframework-camel-case = "^1.3.0"
django-versatileimagefield = "^2.2"
django-simplemde = "^0.1.3"
django-phonenumber-field = {extras = ["phonenumberslite"], version = "^6.0.0"}
django-admin-sso = "^5.0.0"
python-magic-bin = {version = "^0.4.14", platform = "win32"}
django-money = "^2.1.1"
python-dateutil = "^2.8.2"
django-encrypted-model-fields = "^0.6.5"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
yapf = "^0.32.0"
flake8 = "^4.0.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
